changed index file to tsx
